lambda_consumers/ai_consumer/requirements.txt
pydantic>=2.7.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
cryptography>=42.0.0
httpx>=0.27.0
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.7.0
